Revised Movement
Both versions are very choppy, but I believe its because its due to scripts being treated as steps, so when an object is set to move with a script, it will be very stuttery. My second version of the leaving movement supports this; it constantly keeps updating and refreshing the image angle.